# feat(perf-detector-threshold-configuration): Extended consecutive http spans detection.