--- a/src/sentry/utils/performance_issues/detectors/consecutive_http_detector.py
+++ b/src/sentry/utils/performance_issues/detectors/consecutive_http_detector.py
@@ -7,6 +7,10 @@ from sentry.issues.grouptype import PerformanceConsecutiveHTTPQueriesGroupType
 from sentry.issues.issue_occurrence import IssueEvidence
 from sentry.models import Organization, Project
 from sentry.utils.event import is_event_from_browser_javascript_sdk
+from sentry.utils.performance_issues.detectors.utils import (
+    get_max_span_duration,
+    get_total_span_duration,
+)
 from sentry.utils.safe import get_path
 
 from ..base import (
@@ -66,7 +70,7 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
             for span in self.consecutive_http_spans
         )
 
-        exceeds_duration_between_spans_threshold = all(
+        subceeds_duration_between_spans_threshold = all(
             get_duration_between_spans(
                 self.consecutive_http_spans[idx - 1], self.consecutive_http_spans[idx]
             )
@@ -77,7 +81,7 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
         if (
             exceeds_count_threshold
             and exceeds_span_duration_threshold
-            and exceeds_duration_between_spans_threshold
+            and subceeds_duration_between_spans_threshold
         ):
             self._store_performance_problem()
 
@@ -120,13 +124,6 @@ class ConsecutiveHTTPSpanDetector(PerformanceDetector):
         )
 
         self._reset_variables()
-
-    def _sum_span_duration(self, spans: list[Span]) -> float:
-        "Given a list of spans, find the sum of the span durations in milliseconds"
-        sum = 0.0
-        for span in spans:
-            sum += get_span_duration(span).total_seconds() * 1000
-        return sum
 
     def _overlaps_last_span(self, span: Span) -> bool:
         if len(self.consecutive_http_spans) == 0:
@@ -197,7 +194,7 @@ class ConsecutiveHTTPSpanDetectorExtended(ConsecutiveHTTPSpanDetector):
                 "min_time_saved"
             )
 
-        exceeds_duration_between_spans_threshold = all(
+        subceeds_duration_between_spans_threshold = all(
             get_duration_between_spans(
                 self.consecutive_http_spans[idx - 1], self.consecutive_http_spans[idx]
             )
@@ -207,17 +204,14 @@ class ConsecutiveHTTPSpanDetectorExtended(ConsecutiveHTTPSpanDetector):
 
         if (
             exceeds_count_threshold
-            and exceeds_duration_between_spans_threshold
+            and subceeds_duration_between_spans_threshold
             and exceeds_min_time_saved_duration
         ):
             self._store_performance_problem()
 
     def _calculate_time_saved(self) -> float:
-        total_time = self._sum_span_duration(self.consecutive_http_spans)
-
-        max_span_duration = max(
-            [get_span_duration(span).total_seconds() * 1000 for span in self.consecutive_http_spans]
-        )
+        total_time = get_total_span_duration(self.consecutive_http_spans)
+        max_span_duration = get_max_span_duration(self.consecutive_http_spans)
 
         return total_time - max_span_duration