- Added new detector: `ConsecutiveHTTPSpanDetectorExtended`
- Detector only collects metrics. Doesn't create issues. 
- New detector settings:
<img width="450" alt="Screenshot 2023-07-05 at 3 21 16 PM" src="https://github.com/getsentry/sentry/assets/60121741/ddb582f0-228c-4dc6-855e-48ebe858d690">

- We care about time saved from parallelizing for this issue the most. 
- Got rid of individual span duration. Parallelizing 101 spans each taking 10ms , saves 1s. We should create an issue for this. Currently we don't.
- Min time saved handles the misalignment of length case, we don't need min_time_saved_ratio. For example parallelizing 998ms, 1ms, 1ms saves just 2ms which fails the min time saved check. 
- Reduced count threshold to 2, from 3. 
- Added tests of new thresholds and ensured passes all other consecutive http spans tests.